pyproject.toml
[project]
name = "gcri"
version = "0.1.0"
description = "Graph-based Collective Reasoning Interface - A multi-branch reasoning system for LLM agents"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "Dirac-Robot"}
]
keywords = ["llm", "reasoning", "agents", "langgraph", "collective-intelligence", "multi-agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "annotated-types>=0.7.0",
    "anyio>=4.11.0",
    "ato>=2.1.6",
    "certifi>=2025.11.12",
    "charset-normalizer>=3.4.4",
    "distro>=1.9.0",
    "ddgs>=9.10.0",
    "h11>=0.16.0",
    "httpcore>=1.0.9",
    "httpx>=0.28.1",
    "idna>=3.11",
    "jiter>=0.12.0",
    "jsonpatch>=1.33",
    "jsonpointer>=3.0.0",
    "langchain>=1.0.5",
    "langchain-anthropic>=1.2.0",
    "langchain-core>=1.1.0",
    "langchain-google>=0.1.1",
    "langchain-google-genai>=3.2.0",
    "langchain-google-vertexai>=3.1.1",
    "langchain-openai>=1.1.0",
    "langgraph>=1.0.3",
    "langsmith>=0.4.42",
    "loguru>=0.7.3",
    "numpy>=2.3.4",
    "openai>=2.8.1",
    "orjson>=3.11.4",
    "ormsgpack>=1.12.0",
    "packaging>=25.0",
    "pydantic>=2.12.4",
    "pydantic-core>=2.41.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "regex>=2025.11.3",
    "requests>=2.32.5",
    "requests-toolbelt>=1.0.0",
    "rich>=14.2.0",
    "sniffio>=1.3.1",
    "sqlalchemy>=2.0.44",
    "tenacity>=9.1.2",
    "tiktoken>=0.12.0",
    "toml>=0.10.2",
    "typing-extensions>=4.15.0",
    "urllib3>=2.5.0",
    "xxhash>=3.6.0",
    "zstandard>=0.25.0",
    "datasets",
    "inspect-ai>=0.3.38",
    "fastapi",
    "uvicorn",
    "inspect-evals>=0.3.104",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
    "types-requests",
    "types-PyYAML",
]

[project.urls]
Homepage = "https://github.com/Dirac-Robot/GCRI"
Repository = "https://github.com/Dirac-Robot/GCRI"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]

[project.scripts]
gcri = "gcri.entry:main_entry"

[tool.setuptools.package-data]
"gcri" = ["templates/**/*", "dashboard/frontend/dist/**/*"]

[project.entry-points.inspect_ai]
gcri = "gcri.benchmark.model_provider"
